consolidated CredentialItem and ICredentialType
