+++ src/services/api/v4/credential-type.service.ts
@@ -1,5 +1,5 @@
 import { IApiPagedResult } from "@/interfaces/api/result.interface";
-import { ICredentialTypeClaimLabels } from "@/interfaces/api/v4/credential.interface";
+import { ICredentialType } from "@/interfaces/api/v2/credential-type.interface";
 import Http, { HttpResponse } from "@/services/http.service";
 import ApiResource from "@/services/api/resource.service";
 
@@ -9,8 +9,8 @@ export default class CredentialType extends ApiResource {
 
   async getCredentialTypes(
     paging = true
-  ): Promise<HttpResponse<IApiPagedResult<ICredentialTypeClaimLabels>>> {
-    return await Http.get<IApiPagedResult<ICredentialTypeClaimLabels>>(
+  ): Promise<HttpResponse<IApiPagedResult<ICredentialType>>> {
+    return await Http.get<IApiPagedResult<ICredentialType>>(
       this.formatEndpointUrl(""),
       { params: { paging } }
     );
